PublicTransportData & utils.py to separate files
